feat(LadderHeight): 사다리 높이를 나타내는 클래스 작성, 검증 작성
