+++ src/main/java/domain/Ladder.java
@@ -6,13 +6,9 @@ import java.util.List;
 
 public class Ladder {
 
-    private static final int MIN_HEIGHT = 1;
-    private static final int MAX_HEIGHT = 10;
-
     private final List<HorizontalLine> lines = new ArrayList<>();
 
     private Ladder(int playerCount, int height) {
-        validateHeight(height);
         createLadder(playerCount, height);
     }
 
@@ -28,16 +24,6 @@ public class Ladder {
         return lines.stream()
                 .map(HorizontalLine::createStatus)
                 .toList();
-    }
-
-    private void validateHeight(int height) {
-        if (isHeightOutOfRange(height)) {
-            throw new IllegalArgumentException("사다리의 높이는 1 이상 10 이하여야 합니다.");
-        }
-    }
-
-    private boolean isHeightOutOfRange(int height) {
-        return height < MIN_HEIGHT || height > MAX_HEIGHT;
     }
 
     private void createLadder(int playerCount, int height) {
